+++ mkdocs_macros/plugin.py
@@ -18,8 +18,6 @@ from mkdocs.config import config_options
 from mkdocs.config.config_options import Type as PluginType
 from mkdocs.plugins import BasePlugin
 from mkdocs.structure.pages import Page
-from mkdocs.structure.nav import Section
-from mkdocs.utils import get_markdown_title
 
 from mkdocs_macros.errors import format_error
 from mkdocs_macros.context import define_env
@@ -661,20 +659,6 @@ class MacrosPlugin(BasePlugin):
         Capture the nav and files objects so they can be used by
         templates.
         """
-        # Render also the navigation items, so that macros are interpreted
-        # also in navigation
-        # solution to issue #144
-        def render_nav(nav):
-            for nav_item in nav:
-                try:
-                    nav_item.title = self.render(nav_item.title)
-                except AttributeError:
-                    # not title in pre-page navigation, do nothing
-                    pass
-                if isinstance(nav_item, Section): 
-                    # for second, third level titles
-                    render_nav(nav_item.children)
-        render_nav(nav)
         # nav has useful properties like 'pages' and 'items'
         # see: https://github.com/mkdocs/mkdocs/blob/master/mkdocs/structure/nav.py
         self.variables['navigation'] = nav
@@ -738,9 +722,9 @@ class MacrosPlugin(BasePlugin):
                 markdown=self.markdown,
                 # page=page,
             )
-            # HACK: convert macros in the title from render (if exists)
+            # Convert macros in the title from render (if exists)
             # to answer 144
-            page.title = self.render(page.title)
+            page.title = self.render(markdown=page.title)
 
             # execute the post-macro functions in the various modules
             for func in self.post_macro_functions: