+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.1.0-alpha, 2024-04-23
+## 1.0.2, 2023-07-02
+* Added: it is now possible to use macros in page titles, in the
+    nav section of the yaml file, or in the level 1 titles; 
+    the macros are correctly interpreted in the navigation part
+    of the page.
+
+## 1.0.1, 2023-05-25
+
+## 1.0.0-alpha, 2023-04-23
 
 * Improved user guide, with introduction of two new pages:
     "Controlling macro rendering" and "Post production".